Display a dialouge before deleting

+++ FriendOrganizer.UI/ViewModel/FriendDetailViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using FriendOrganizer.Model;
 using FriendOrganizer.UI.Data.Repositories;
 using FriendOrganizer.UI.Event;
+using FriendOrganizer.UI.View.Services;
 using FriendOrganizer.UI.Wrapper;
 using Prism.Commands;
 using Prism.Events;
@@ -16,6 +17,8 @@ namespace FriendOrganizer.UI.ViewModel
         private FriendWrapper _friend;
         private IFriendRepository _friendRepository;
         private bool _hasChanges;
+        private IMessageDialogService _messageDialogService;
+
 
         private Friend CreateNewFriend()
         {
@@ -36,10 +39,11 @@ namespace FriendOrganizer.UI.ViewModel
             }
         }
 
-        public FriendDetailViewModel(IFriendRepository friendRepository, IEventAggregator eventAggregator)
+        public FriendDetailViewModel(IFriendRepository friendRepository, IEventAggregator eventAggregator, IMessageDialogService messageDialogService)
         {
             _friendRepository = friendRepository;
             _eventAggregator = eventAggregator;
+            _messageDialogService = messageDialogService;
 
             SaveCommand = new DelegateCommand(OnSaveExecute, OnSaveCanExecute);
             DeleteCommand = new DelegateCommand(OnDeleteExecute);
@@ -47,9 +51,13 @@ namespace FriendOrganizer.UI.ViewModel
 
         private async void OnDeleteExecute()
         {
-            _friendRepository.Remove(Friend.Model);
-            await _friendRepository.SaveAsync();
-            _eventAggregator.GetEvent<AfterFriendDeletedEvent>().Publish(Friend.Id);
+            var result = _messageDialogService.ShowOkCancelDialog($"Do you really want to delete the {Friend.FirstName} {Friend.LastName}", "Question");
+            if (result == MessageDialogResult.OK) 
+            {
+                _friendRepository.Remove(Friend.Model);
+                await _friendRepository.SaveAsync();
+                _eventAggregator.GetEvent<AfterFriendDeletedEvent>().Publish(Friend.Id);
+            }
         }
 
         public bool HasChanges
